+ added not existing check on Carlistings/Customers Delete

+++ src/main/java/com/avd/jdmrest/controller/CustomerController.java
@@ -2,6 +2,7 @@ package com.avd.jdmrest.controller;
 
 import com.avd.jdmrest.domain.Customer;
 import com.avd.jdmrest.services.CustomerService;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.Optional;
@@ -32,7 +33,12 @@ public class CustomerController {
 	}
 
 	@DeleteMapping("/{id}")
-	public void deleteCustomer(@PathVariable Long id) {
-		customerService.deleteById(id);
+	public ResponseEntity<String> deleteCustomer(@PathVariable Long id) {
+		if (customerService.getById(id).isPresent()) {
+			customerService.deleteById(id);
+			return ResponseEntity.ok("Customer deleted");
+		} else {
+			return ResponseEntity.notFound().build();
+		}
 	}
 }
